pyproject.toml
[project]
name = "ai-inspection"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "accelerate==1.12.0",
    "fastapi>=0.129.0",
    "matplotlib>=3.10.8",
    "numpy==2.2.3",
    "opencv-python==4.13.0.92",
    "pillow==12.1.1",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "safetensors==0.7.0",
    "torch==2.9.1",
    "torchaudio==2.9.1",
    "torchvision==0.24.1",
    "transformers==5.2.0",
    "uvicorn>=0.41.0",
]
